Create dialect-specific column types
