+++ packages/schemix/src/schemix/postgres/__init__.py
@@ -1,0 +1,41 @@
+"""PostgreSQL dialect for Schemix ORM."""
+
+from .columns import (
+    JSON,
+    JSONB,
+    BigInt,
+    BigSerial,
+    Boolean,
+    Char,
+    Date,
+    Decimal,
+    Integer,
+    Numeric,
+    Serial,
+    SmallInt,
+    SmallSerial,
+    Text,
+    Time,
+    Timestamp,
+    Varchar,
+)
+
+__all__ = [
+    "Integer",
+    "SmallInt",
+    "BigInt",
+    "Serial",
+    "SmallSerial",
+    "BigSerial",
+    "Numeric",
+    "Decimal",
+    "Varchar",
+    "Text",
+    "Char",
+    "Boolean",
+    "Date",
+    "Time",
+    "Timestamp",
+    "JSON",
+    "JSONB",
+]
